refactor: :truck: rename CamelCase to snake-case of component folder name
