+++ src/components/blog/explorer-section/open-post-list/open-post-list-item/index.style.ts
