+++ src/components/blog/open-post-tabs/index.style.ts
