+++ src/components/blog/open-post-tabs/open-post-tab/index.style.ts
