+++ src/components/blog/post-template/index.style.ts
