+++ src/components/common/layouts/index.ts
@@ -1,1 +1,1 @@
-export { default as MainLayout } from "./MainLayout";
+export { default as MainLayout } from "./main-layout";
