+++ src/components/common/layouts/main-layout/activity-bar/action-item/index.style.ts
