+++ src/components/common/layouts/main-layout/activity-bar/index.style.ts
