+++ src/components/common/layouts/main-layout/index.style.ts
