+++ src/components/common/molecules/index.ts
@@ -1,4 +1,2 @@
-export * from "./AccordionSection";
-export { default as AccordionSection } from "./AccordionSection";
-
-export { default as ThemeSwitcher } from "./ThemeSwitcher";
+export { default as AccordionSection } from "./accordion-section";
+export { default as ThemeSwitcher } from "./theme-switcher";
