+++ src/components/common/organisms/index.ts
@@ -1,1 +1,1 @@
-export { default as FileTree } from "./FileTree";
+export { default as FileTree } from "./file-tree";
